+++ jablib/src/main/java/org/jabref/model/entry/field/FieldProperty.java
@@ -35,5 +35,6 @@ public enum FieldProperty {
 
     // Field content should be treated as data
     VERBATIM,
+
     YES_NO
 }